+++ lib/src/clixon_uid.c
@@ -55,7 +55,10 @@
 #include <grp.h>
 #include <pwd.h>
 
-/* clicon */
+/* cligen */
+#include <cligen/cligen.h>
+
+/* clixon */
 #include "clixon_err.h"
 #include "clixon_log.h"
 #include "clixon_uid.h"